History is working! Add try/catch to home page. Remove delete user from main.

+++ src/main/java/com/app/BootConverter/entities/Currency.java
@@ -22,12 +22,12 @@ public class Currency {
 	String charCode;
 
 	@Column(name = "nominal")
-	Integer nominal;
+	Long nominal;
 
 	@Column(name = "value")
 	Double value;
 
-	public Currency(String name, String charCode, Integer nominal, Double value) {
+	public Currency(String name, String charCode, Long nominal, Double value) {
 		this.name = name;
 		this.charCode = charCode;
 		this.nominal = nominal;
