+++ src/main/java/com/app/BootConverter/services/CurrencyConverterService.java
@@ -9,7 +9,7 @@ public interface CurrencyConverterService {
 
 	String getName(String id);
 
-	Integer getNominal(Long id);
+	Long getNominal(Long id);
 
 	Double getValue(Long id);
